% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Node.R
\docType{methods}
\name{archive}
\alias{archive}
\alias{archive,D1Node,character-method}
\title{Archive an object on a Member Node or Coordinating Node, which hides it from casual searches.}
\usage{
archive(node, pid, ...)

\S4method{archive}{D1Node,character}(node, pid, quiet = TRUE)
}
\arguments{
\item{node}{The MNode or CNode instance on which the object will be created}

\item{pid}{The identifier of the object to be created}

\item{...}{(Not yet used)}

\item{quiet}{A logical value - if TRUE (the default) then informational messages are not printed.}
}
\value{
The pid that was archived if successful, otherwise NULL
}
\description{
This method provides the ability to archive a data or metadata object on the Member Node
provided in the \code{'mnode'} parameter.  Archiving removes the object from DataONE search functions,
thereby making it more difficult to find without completely removing the object.  Archive is intended
for objects that should not be used by current researchers, but for which there is a desire to maintain
a historical record, such as when journal articles might cite the object.  Users can still obtain the
contents of archived objects if they have the identifier, but will not discover it through searches.
}
\details{
This operation requires an X.509 certificate to be present in the default location of the file
system. This certificate provides authentication credentials from
CILogon \url{https://cilogon.org/?skin=DataONE}.  See \code{\link{{CertificateManager}}} for details.
Also, administrator priviledge is required to run archive() on a DataONE Coordinating Node.
}
\section{Methods (by class)}{
\itemize{
\item \code{node = D1Node,pid = character}: 
}}

