% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Object.R
\docType{methods}
\name{asDataFrame,D1Object,D1Object-method}
\alias{asDataFrame,D1Object,D1Object-method}
\title{this method uses the provided metadata reference object for instructions on
how to parse the data table (which parameters to set)
'reference' is the metadata D1Object that gives instruction on how to read the data
into the dataFrame}
\usage{
\S4method{asDataFrame}{D1Object,D1Object}(x, reference, ...)
}
\description{
this method uses the provided metadata reference object for instructions on
how to parse the data table (which parameters to set)
'reference' is the metadata D1Object that gives instruction on how to read the data
into the dataFrame
}

