% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auth_request.R
\name{auth_delete}
\alias{auth_delete}
\title{DELETE a resource with authenticated credentials.}
\usage{
auth_delete(url, encode = "multipart", body = as.list(NA))
}
\arguments{
\item{url}{The URL to be accessed via authenticated DELETE}

\item{encode}{the type of encoding to use for the DELETE body, defaults to 'multipart'}

\item{body}{a list of data to be included in the body of the DELETE request}
}
\value{
the HTTP response from the request
}
\description{
DELETE data at a URL using an HTTP DELETE request using authentication credentials
provided in a client certificate.  Authenticated access depends on the suggested
PKIplus package. If the PKIplus package is not installed, then the request fails.
}

