% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auth_request.R
\name{auth_put}
\alias{auth_put}
\title{PUT a resource with authenticated credentials.}
\usage{
auth_put(url, encode = "multipart", body = as.list(NA), node = NULL)
}
\arguments{
\item{url}{The URL to be accessed via authenticated PUT}

\item{encode}{the type of encoding to use for the PUT body, defaults to 'multipart'}

\item{body}{a list of data to be included in the body of the PUT request}
}
\value{
the HTTP response from the request
}
\description{
PUT data to a URL using an HTTP PUT request using authentication credentials
provided in a client certificate.  Authenticated access depends on the suggested
PKIplus package. If the PKIplus package is not installed, then the request fails.
}

