% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auth_request.R
\name{auth_put_post_delete}
\alias{auth_put_post_delete}
\title{POST, PUT, or DELETE a resource with authenticated credentials.}
\usage{
auth_put_post_delete(method, url, encode = "multipart", body = as.list(NA))
}
\arguments{
\item{method}{a string indicating which HTTP method to use (post, put, or delete)}

\item{url}{The URL to be accessed via authenticated PUT}

\item{encode}{the type of encoding to use for the PUT body, defaults to 'multipart'}

\item{body}{a list of data to be included in the body of the PUT request}
}
\value{
the response object from the method
}
\description{
POST, PUT, or DELETE data to a URL using an HTTP request using authentication credentials
provided in a client certificate.  Authenticated access depends on the suggested
PKIplus package. If the PKIplus package is not installed, then the request fails.
}

