% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AuthenticationManager.R
\docType{class}
\name{AuthenticationManager-class}
\alias{AuthenticationManager,ANY-method}
\alias{AuthenticationManager-class}
\alias{getAuthExpires,AuthenticationManager-method}
\title{AuthenticationManager provides mechanisms to}
\usage{
\S4method{AuthenticationManager}{ANY}()

\S4method{getAuthExpires}{AuthenticationManager}(x)
}
\description{
AuthenticationManager
}
\details{
Understanding how your identity is managed is important for working with DataONE, especially to
avoid unexpected results. For example, depending your authorization status, searches may or may
return only public records, or the full set of public and private records. Object and package
retrievals might fail if some or all of the objects being retrieved are private.  Creating or
updating objects on DataONE nodes and reserving identifiers reservations might fail if your
authorization certificate is missing or expired.

DataONE identifies you using CILogon-provided x509 certificates. DataONE has
partnered with CILogon to provide a widely-accessible certificate issuing mechanism
that allows DataONE users to use existing trusted institutional and public accounts.

DataONE version 2.0.0 provides an addition authentication mechanism known as
authentication tokens. For information about tokens and instructions for generating
a token for use with the DataONE R client, view the package overview by
entering the command: \code{'vignette("dataone-overview")'}

CILogon recognizes many identity providers, including many universities as well as
Google, so most times users new to DataONE can get certificates using one
of their existing accounts. For more information about the CILogon service, see
\url{https://cilogon.org/?skin=DataONE} .
}
\section{Methods (by generic)}{
\itemize{
\item \code{AuthenticationManager}: 

\item \code{getAuthExpires}: 
}}
\section{Slots}{

\describe{
\item{\code{authInfo}}{value of type \code{"hash"} containing authentication information.}
}}
\examples{
am <- AuthenticationManager()
}

