% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AuthenticationManager.R
\name{AuthenticationManager}
\alias{AuthenticationManager}
\title{Create an AuthenticationManager object}
\usage{
AuthenticationManager(...)
}
\value{
the AuthenticationManager object
}
\description{
Construct an instance of AuthenticationManager to provide mechanisms to load, verify, and
display DataONE authenticatin information.
}

