% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Object.R
\docType{methods}
\name{canRead,D1Object,character-method}
\alias{canRead,D1Object,character-method}
\title{Test whether the provided subject can read the object}
\usage{
\S4method{canRead}{D1Object,character}(x, subject)
}
\arguments{
\item{x}{D1Client}

\item{subject}{: character}

\item{...}{(not yet used)}
}
\value{
TRUE or FALSE
}
\description{
Using the AccessPolicy, tests whether the subject has read permission
for the object.  This method is meant work prior to submission, so uses
only the AccessPolicy to determine who can read (Not the rightsHolder field,
which always can read.)
}
\author{
rnahf
}

