% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CNode.R, R/D1Node.R
\docType{methods}
\name{listFormats,CNode-method}
\alias{describe,CNode,character-method}
\alias{get,CNode,character-method}
\alias{getChecksum,CNode,character-method}
\alias{getFormat}
\alias{getFormat,CNode-method}
\alias{getQueryEngineDescription,D1Node,character-method}
\alias{getSystemMetadata,CNode,character-method}
\alias{listFormats}
\alias{listFormats,CNode-method}
\alias{listQueryEngines,D1Node-method}
\title{list formats}
\usage{
\S4method{listFormats}{CNode}(cnode)

\S4method{getFormat}{CNode}(cnode, formatId)

\S4method{getChecksum}{CNode,character}(node, pid)

\S4method{get}{CNode,character}(node, pid)

\S4method{getSystemMetadata}{CNode,character}(node, pid)

\S4method{describe}{CNode,character}(node, pid)

\S4method{getQueryEngineDescription}{D1Node,character}(node, queryEngineName)

\S4method{listQueryEngines}{D1Node}(node)
}
\arguments{
\item{node}{The CNode instance from which the checksum will be retrieved}

\item{pid}{The identifier of the object}

\item{node}{The CNode instance from which the pid will be downloaded}

\item{pid}{The identifier of the object to be downloaded}

\item{node}{The CNode instance from which the SystemMetadata will be downloaded}

\item{pid}{The identifier of the object}

\item{mnode}{The CNode instance from which the identifier will be generated}

\item{pid}{Identifier for the object in question. May be either a PID or a SID. Transmitted as
part of the URL path and must be escaped accordingly.}

\item{node}{The CNode or MNode instance to list the query engines for.}

\item{node}{The CNode or MNode instance to list the query engines for.}
}
\value{
character the checksum value, with the checksum algorithm as the attribute "algorithm"

the bytes of the object

SystemMetadata for the object

A list of header elements

list Objects that met the search criteria

list Objects that met the search criteria
}
\description{
list formats

A checksum is calculated for an object when it is uploaded to DataONE and
is submitted with the object's system metadata. The \code{'getChecksum'} method retrieves
the checksum from the specified coordinating node

Get the bytes associated with an object on this Coordinating Node.

The SystemMetadata includes information about the identity, type, access control, and other system
level details about the object.

This method provides a lighter weight mechanism than getSystemMetadata() for a client to
determine basic properties of the referenced object.

Query a node for the list of query engines available on the node

Query a node for the list of query engines available on the node
}
\details{
This operation acts as the 'public' anonymous user unless an X.509 certificate is
present in the default location of the file system, in which case the access will be authenticated.

This operation acts as the 'public' anonymous user unless an X.509 certificate is
present in the default location of the file system, in which case the access will be authenticated.
}
\section{Methods (by generic)}{
\itemize{
\item \code{listFormats}: 

\item \code{getFormat}: 

\item \code{getChecksum}: 

\item \code{get}: 

\item \code{getSystemMetadata}: 

\item \code{describe}: 

\item \code{getQueryEngineDescription}: 

\item \code{listQueryEngines}: 
}}
\examples{
\dontrun{
cn <- CNode("PROD")
engineDesc <- getQueryEngineDescription(cn, "solr")
cat(sprintf("Query engine version: \%s\\n", engineDesc[1]$queryEngineVersion))
cat(sprintf("Query engine name: \%s\\n", engineDesc[2]$name))
for (i in 5:length(engineDesc)) {
  cat(sprintf("query field: \%s : \%s\\n", engineDesc[i]$queryField$name, engineDesc[i]$queryField$description))
}
}
}
\author{
Scott Chamberlain
}
\seealso{
\url{http://mule1.dataone.org/ArchitectureDocs-current/apis/CN_APIs.html#CNRead.getChecksum}

\url{http://mule1.dataone.org/ArchitectureDocs-current/apis/CN_APIs.html#CNRead.get}

\url{http://mule1.dataone.org/ArchitectureDocs-current/apis/CN_APIs.html#CNRead.getSystemMetadata}

http://mule1.dataone.org/ArchitectureDocs-current/apis/CN_APIs.html#CNRead.describe
}

