% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Client.R
\docType{methods}
\name{convert.csv}
\alias{convert.csv}
\alias{convert.csv,D1Client-method}
\title{Convert a DataFrame to Standard CSV.}
\usage{
convert.csv(x, ...)

\S4method{convert.csv}{D1Client}(x, df, ...)
}
\arguments{
\item{x}{A D1Client object}

\item{...}{additional params passed to write.csv}

\item{df}{the dataFrame}
}
\value{
the dataframe serialized as a .csv
}
\description{
Convert a DataFrame to Standard CSV.
}
\section{Methods (by class)}{
\itemize{
\item \code{D1Client}: 
}}
\seealso{
\code{\link[=D1Client-class]{D1Client}}{ class description.}
}

