% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Client.R
\docType{methods}
\name{createD1Object}
\alias{createD1Object}
\alias{createD1Object,D1Client,D1Object-method}
\title{Create the Object in the DataONE System}
\usage{
createD1Object(x, d1Object, ...)

\S4method{createD1Object}{D1Client,D1Object}(x, d1Object)
}
\arguments{
\item{x}{: D1Client}

\item{d1Object}{A D1Object instance to upload to DataONE}

\item{...}{(not yet used)}
}
\value{
TRUE if the object was successfully uploaded, FALSE if not.
}
\description{
Create the Object in the DataONE System
}
\section{Methods (by class)}{
\itemize{
\item \code{x = D1Client,d1Object = D1Object}: 
}}
\examples{
\dontrun{
d1c <- D1Client(env="STAGING", mNodeid="urn:node:mnStageUCSB2")
d1o <- new("D1Object", file=system.file("extdata/sample-eml.xml", package="dataone"))
d1o <- setPublicAccess(d1o)
uploaded <- createD1Object(d1c, d1o)
}
}

