% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/D1Client.R
\docType{methods}
\name{D1Client,ANY,ANY-method}
\alias{D1Client,ANY,ANY-method}
\title{Construct a D1Client, using default env ("PROD") and nodeid ("")}
\usage{
\S4method{D1Client}{ANY,ANY}()
}
\value{
the D1Client object representing the DataONE environment
}
\description{
Construct a D1Client, using default env ("PROD") and nodeid ("")
}

