% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{D1Client,character,ANY-method}
\alias{D1Client,character,ANY-method}
\title{Pass in the environment to be used by this D1Client, but use
  the default member node.}
\usage{
\S4method{D1Client}{character,ANY}(env, mnNodeid, ...)
}
\description{
Pass in the environment to be used by this D1Client, but use
  the default member node.
}

