% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Client.R
\docType{methods}
\name{D1Client,character,character-method}
\alias{D1Client,character,character-method}
\title{Pass in the environment to be used by this D1Client, plus the
id of the member node to be used for primary interactions such as creates}
\usage{
\S4method{D1Client}{character,character}(env, mnNodeid)
}
\description{
Pass in the environment to be used by this D1Client, plus the
id of the member node to be used for primary interactions such as creates
}

