% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Client.R
\docType{class}
\name{D1Client-class}
\alias{D1Client-class}
\title{The D1Client class contains methods that perform high level DataONE tasks}
\description{
The methods in the D1Client class call the low level DataONE API to
perform involved tasks such as uploading all the packages in a DataPackage (i.e
\code{\link{uploadDataPackage}})
}
\section{Slots}{

\describe{
\item{\code{endpoint}}{The baseurl of the CN in that en  vironment}

\item{\code{mn.nodeid}}{The NodeReference for the 'home' MemberNode for this application, where creates/updates will happen.}

\item{\code{client}}{The reference to the internally held Java D1Client instance}

\item{\code{session}}{The reference to the internally held Java Session instance}
}}
\section{Methods}{

\itemize{
 \item{\code{\link[=initialize-D1Client]{initialize}}}{: Initialize a Recordr object}
 \item{\code{\link{getPackage}}}{: Begin recording provenance for an R session}
 \item{\code{\link{getD1Object}}}{: Get the Identifiers of Package Members}
 \item{\code{\link{getDataObject}}}{: Get the data content of a specified data object}
 \item{\code{\link{d1SolrQuery}}}{: Add a DataObject to the DataPackage}
 \item{\code{\link{d1IdentifierSearch}}}{: Record relationships of objects in a DataPackage}
 \item{\code{\link{reserveIdentifier}}}{: Record derivation relationships between objects in a DataPackage}
 \item{\code{\link{createDataPackage}}}{: Retrieve relationships of package objects}
 \item{\code{\link{getEndpoint}}}{: Retrieve relationships of package objects}
 \item{\code{\link{getMNodeId}}}{: Retrieve relationships of package objects}
 \item{\code{\link{getMN}}}{: Retrieve relationships of package objects}
 \item{\code{\link{uploadDataPackage}}}{: Retrieve relationships of package objects}
 \item{\code{\link{uploadDataObject}}}{: Retrieve relationships of package objects}
 \item{\code{\link{listMemberNodes}}}{: Retrieve relationships of package objects}
 \item{\code{\link{convert.csv}}}{: Retrieve relationships of package objects}
 \item{\code{\link{encodeUrlQuery}}}{: Retrieve relationships of package objects}
 \item{\code{\link{encodeUrlPath}}}{: Retrieve relationships of package objects}
}
}
\seealso{
\code{\link{dataone}}{ package description.}
}

