% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Client.R
\docType{methods}
\name{D1Client}
\alias{D1Client}
\alias{D1Client,ANY,ANY-method}
\alias{D1Client,character,ANY-method}
\alias{D1Client,character,character-method}
\alias{D1Client-initialize}
\alias{initialize,D1Client-method}
\title{The DataONE client class used to downlaod, update and search for data in the DataONE network.}
\usage{
D1Client(env, mNodeid, ...)

\S4method{D1Client}{ANY,ANY}()

\S4method{D1Client}{character,ANY}(env, mNodeid, ...)

\S4method{D1Client}{character,character}(env, mNodeid)

\S4method{initialize}{D1Client}(.Object, cn = NA, mn = NA,
  env = as.character(NA), mNodeid = as.character(NA))
}
\arguments{
\item{env}{The label for the DataONE environment to be using ('PROD','STAGING','SANDBOX','DEV')}

\item{...}{(not yet used)}

\item{mNodeidThe}{node Id of the application's 'home' node.  Should be already registered to the corresponding 'env'}
}
\value{
the D1Client object representing the DataONE environment

the D1Client object representing the DataONE environment
}
\description{
The DataONE client class used to downlaod, update and search for data in the DataONE network.

Construct a D1Client, using default env ("PROD") and nodeid ("")

Pass in the environment to be used by this D1Client, but use

Pass in the environment to be used by this D1Client, plus the

Initialize a D1Client object
}
\author{
mbjones
}
\seealso{
\code{\link[=D1Client-class]{D1Client}}{ class description.}

\code{\link[=D1Client-class]{D1Client}}{ class description.}

\code{\link[=D1Client-class]{D1Client}}{ class description.}

\code{\link[=D1Client-class]{dataone}}{ class description.}
}

