% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Node.R
\docType{class}
\name{D1Node-class}
\alias{D1Node,XMLInternalElementNode-method}
\alias{D1Node-class}
\title{A class representing a Member Node repository, which can expose and store data}
\usage{
\S4method{D1Node}{XMLInternalElementNode}(xml)
}
\description{
A class representing a Member Node repository, which can expose and store data
}
\section{Methods (by generic)}{
\itemize{
\item \code{D1Node}: 
}}
\section{Slots}{

\describe{
\item{\code{identifier}}{The node identifier of the MN}

\item{\code{name}}{The node name}

\item{\code{description}}{The node description}

\item{\code{baseURL}}{The registered baseURL for the node, which does not include the version string}

\item{\code{subject}}{The Distinguished Name of this node, used for authentication}

\item{\code{contactSubject}}{The Distinguished Name of contact person for this node}

\item{\code{replicate}}{a boolean flag indicating whether the node accepts replicas}

\item{\code{type}}{the node type, either 'mn' or 'cn'}

\item{\code{state}}{an indication of whether the node is accessible, either 'up' or 'down'}

\item{\code{services}}{A data.frame containing the service tiers supported by this node.}

\item{\code{serviceUrls}}{a data.frame that contains DataONE service Urls}

\item{\code{APIversion}}{The version of the DataONE API for this node}
}}

