% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Node.R
\name{D1Node}
\alias{D1Node}
\title{Create a D1Node object.}
\usage{
D1Node(xml, ...)
}
\arguments{
\item{xml}{An XML object that describes the node to be initialized (see \link{listNodes}).}

\item{...}{(not yet used)}
}
\value{
the Node object representing the DataONE environment
}
\description{
Create a D1Node object.
}

