% Generated by roxygen2 (4.0.2): do not edit by hand
\name{d1SolrQuery}
\alias{d1SolrQuery}
\title{A method to query the DataONE solr endpoint of the Coordinating Node.
It expects any lucene reserved characters to already be escaped with backslash. If
solrQuery is a list, it is expected to have field names as attributes and search
values as the values in the list.}
\usage{
d1SolrQuery(x, solrQuery)
}
\arguments{
\item{x}{the D1Client (environment) being queried}

\item{solrQuery}{list or character: a fully encoded query string}
}
\value{
the solr response (XML)
}
\description{
A method to query the DataONE solr endpoint of the Coordinating Node.
It expects any lucene reserved characters to already be escaped with backslash. If
solrQuery is a list, it is expected to have field names as attributes and search
values as the values in the list.
}
\examples{
\dontrun{ d1SolrQuery(client,list(q="+species population diversity", fl="identifier")) }
}

