% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractTableDescriber.R, R/EMLParser.R
\docType{methods}
\name{data.formatFamily}
\alias{data.formatFamily}
\alias{data.formatFamily,EMLParser,numeric-method}
\title{Data Format}
\usage{
data.formatFamily(x, index, ...)

\S4method{data.formatFamily}{EMLParser,numeric}(x, index)
}
\arguments{
\item{x}{the TableDescriber}

\item{index}{index of the table within the document}

\item{...}{Additional parameters}
}
\value{
the format of the data object being described
}
\description{
Get the table format family.
}
\author{
rnahf
}

