% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataone-package.R
\docType{package}
\name{dataone}
\alias{dataone}
\alias{dataone-package}
\title{Search, download and upload data to the DataONE network.}
\description{
The R package \emph{dataone} provides read/write access to data and metadata from the DataONE network of
Member Node data repositories. Member Nodes in DataONE are independent data repositories that have adopted the DataONE
services for interoperability, making each of the repositories accessible to client tools such as the DataONE R Client
using a standard interface. The DataONE R Client can be used to access data files and to write new data and metadata files
to nodes in the DataONE network.

A description of the \emph{dataone} R package is available with the command: \code{'vignette("dataone-overview")'}.
}
\author{
Matthew B. Jones (NCEAS) and Peter Slaughter (NCEAS)
}

