% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CNode.R, R/D1Node.R
\docType{methods}
\name{describe,CNode,character-method}
\alias{describe}
\alias{describe,CNode,character-method}
\title{Efficiently determine basic properties about the requested object.}
\usage{
\S4method{describe}{CNode,character}(node, pid)

describe(node, pid, ...)
}
\arguments{
\item{pid}{Identifier for the object in question. May be either a PID or a SID. Transmitted as
part of the URL path and must be escaped accordingly.}

\item{mnode}{The CNode instance from which the identifier will be generated}

\item{mnode}{The Node instance from which the identifier will be generated}

\item{pid}{Identifier for the object in question. May be either a PID or a SID. Transmitted as
part of the URL path and must be escaped accordingly.}
}
\value{
A list of header elements

A list of header elements
}
\description{
This method provides a lighter weight mechanism than getSystemMetadata() for a client to
determine basic properties of the referenced object.

This method provides a lighter weight mechanism than getSystemMetadata() for a client to
determine basic properties of the referenced object.
}
\seealso{
\code{\link[=CNode-class]{CNode}}{ class description.}
}

