% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractTableDescriber.R, R/EMLParser.R
\docType{methods}
\name{documented.sizes}
\alias{documented.sizes}
\alias{documented.sizes,EMLParser-method}
\title{Get the sizes of the described data tables.}
\usage{
documented.sizes(x, ...)

\S4method{documented.sizes}{EMLParser}(x)
}
\arguments{
\item{x}{the TableDescriber}

\item{...}{Additional parameters}
}
\value{
vector of data table sizes (in bytes)
}
\description{
Get the table size.
}
\author{
rnahf
}

