% Generated by roxygen2 (4.0.2): do not edit by hand
\name{downloadCert}
\alias{downloadCert}
\title{Open the CILogon Certificate download page in the default browser.}
\usage{
downloadCert(x, ...)
}
\arguments{
\item{x}{a CertificateManager instance}
}
\description{
A convenience method to take you to the CILogon download page:
\url{'https://cilogon.org/?skin=DataONE'}. Logging into CILogon will allow
you to download your X.509 certificate to your local computer.  Typically,
the certificate is saved in the default Globus location for certificates
(\code{\link{showCertLocation}}) and once it is there, the \code{'dataone'}
package will use the certificate for all authenticated operations.  Deleting
the certificate file is the equivalent of logging out.
}

