% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMLParser.R
\docType{class}
\name{EMLParser-class}
\alias{EMLParser-class}
\title{Handler for Parsing Table Format Details from Metadata}
\description{
#' Implements methods to provide parsing instructions for asDataFrame.
}
\details{
#' handles eml formats 2.0.0 through 2.1.1
}
\section{Slots}{

\describe{
\item{\code{d1Object}}{the metadata object}

\item{\code{xmlDocRoot}}{the xml representation of the metadata}
}}
\author{
rnahf
}

