% Generated by roxygen2 (4.0.2): do not edit by hand
\name{encodeSolr}
\alias{encodeSolr}
\title{Encode the Input for Solr Queries}
\usage{
encodeSolr(x, segment, ...)
}
\arguments{
\item{segment}{: a string to encode}
}
\value{
the encoded form of the input
}
\description{
Treating all special characters and spaces as literals, backslash escape special
characters, and double-quote if necessary
}
\examples{
encodeSolr("this & that")
}

