% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CNode.R
\name{encodeSolrQuery}
\alias{encodeSolrQuery}
\title{Encode input for a Url Query segment}
\usage{
encodeSolrQuery(querySegment, ...)
}
\arguments{
\item{query:}{a string to encode}
}
\value{
the encoded form of the input
}
\description{
Encodes the characters of the input so they are not interpretted as reserved
characters.  Will also encode non-ASCII unicode characters.
}
\examples{
fullyEncodedQuery <- paste0("q=id:",encodeUrlQuery(client,encodeSolr("doi:10.6085/AA/YBHX00_XXXITBDXMMR01_20040720.50.5")))
}

