% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/D1Client.R
\name{encodeUrlPath}
\alias{encodeUrlPath}
\title{Encode the Input for a URL Path Segment}
\usage{
encodeUrlPath(x, pathSegment, ...)
}
\arguments{
\item{pathSegment}{: a string to encode}
}
\value{
the encoded form of the input
}
\description{
Encodes the characters of the input so they are not interpretted as reserved
characters in url strings.  Will also encode non-ASCII unicode characters.
}
\examples{
fullyEncodedPath <- paste0("cn/v1/object/",encodeUrlPath("doi:10.6085/AA/YBHX00_XXXITBDXMMR01_20040720.50.5"))
}

