% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MNode.R
\name{generateIdentifier}
\alias{generateIdentifier}
\title{Get a unique identifier that is generated by the Member Node repository and guaranteed to be unique.}
\usage{
generateIdentifier(mnode, ...)
}
\arguments{
\item{mnode}{The MNode instance on which the object will be created}

\item{scheme}{The identifier scheme to be used, such as DOI, UUID, etc.}

\item{fragment}{An optional fragment to be prepended to the identifier for schemes that support it (not all do).}
}
\value{
the character string of the generated unique identifier
}
\description{
Creating objects requires use of a unique persistent identifier (pid) when calling the create
function.  Member Nodes may optionally provide the generateIdentifier service to issue such identifiers,
ensuring that they are unique. Each identifier conforms to an identifier scheme, which determines the syntax and
rules for how the identifier that is generated is formatted.  All Member Nodes that implement this method must
support the UUID scheme, but may also support other schemes such as DOI and others.
}
\details{
This operation requires an X.509 certificate to be present in the default location of the file
system. This certificate provides authentication credentials from
CILogon \url{https://cilogon.org/?skin=DataONE}.  See \code{\link{{CertificateManager}}} for details.
}
\seealso{
\url{http://mule1.dataone.org/ArchitectureDocs-current/apis/MN_APIs.html#MNStorage.generateIdentifier}
}

