% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AuthenticationManager.R
\docType{methods}
\name{getAuthMethod}
\alias{getAuthMethod}
\alias{getAuthMethod,AuthenticationManager-method}
\title{Get the current valid authentication mechanism.}
\usage{
getAuthMethod(x, ...)

\S4method{getAuthMethod}{AuthenticationManager}(x)
}
\arguments{
\item{x}{A AuthenticationManager instance}

\item{...}{(Not yet used)}
}
\value{
The current authentication mechanism as a character string, either "token" or "cert".
}
\description{
Get the current valid authentication mechanism.
}
\section{Methods (by class)}{
\itemize{
\item \code{AuthenticationManager}: 
}}

