% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthenticationManager.R
\docType{methods}
\name{getAuthSubject}
\alias{getAuthSubject}
\alias{getAuthSubject,AuthenticationManager-method}
\title{Get the authentication subject.}
\usage{
getAuthSubject(x, ...)

\S4method{getAuthSubject}{AuthenticationManager}(x)
}
\arguments{
\item{x}{an AuthenticationManager instance}

\item{...}{(Not yet used)}
}
\value{
the DataONE Subject that is your client's identity
}
\description{
Get the authentication subject.
}

