% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MNode.R
\name{getCapabilities}
\alias{getCapabilities}
\title{Get the node capabilities description, and store the information in the MNode.}
\usage{
getCapabilities(mnode, ...)
}
\arguments{
\item{mnode}{The node identifier with which this node is registered in DataONE}
}
\value{
the MNode object representing the DataONE environment
}
\description{
Access the DataONE getCapabilities() service for the Member Node, which returns an XML
decription of the repository and the services it offers.
}
\seealso{
\url{http://mule1.dataone.org/ArchitectureDocs-current/apis/MN_APIs.html#MN_core.getCapabilities}
}

