% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getCertExpires}
\alias{getCertExpires}
\title{Show the date and time when an X.509 certificate expires.}
\usage{
getCertExpires(x, ...)
}
\arguments{
\item{x}{a CertificateManager instance}
}
\value{
POSIXct value
}
\description{
Each X.509 has a range of certificate validity times.  This method returns the X.509
\code{'notAfter'} field formatted as a \code{'POSIXct'} date value.
}

