% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Node.R
\name{getChecksum}
\alias{getChecksum}
\title{Returns the checksum for a pid}
\usage{
getChecksum(node, pid, ...)
}
\arguments{
\item{node}{The CNode or MNode instance from which the checksum will be retrieved}

\item{pid}{The object identifier to be downloaded}
}
\value{
checksum The comuted hash of the pid object
}
\description{
Returns the checksum for a pid
}

