% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Client.R
\docType{methods}
\name{getCN}
\alias{getCN}
\alias{getCN,D1Client-method}
\title{Get the coordinating node associated with this D1Client object.}
\usage{
getCN(x)

\S4method{getCN}{D1Client}(x)
}
\arguments{
\item{x}{A D1Client object.}
}
\description{
Get the coordinating node associated with this D1Client object.
}
\section{Methods (by class)}{
\itemize{
\item \code{D1Client}: 
}}
\seealso{
\code{\link[=D1Client-class]{D1Client}}{ class description.}
}

