% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SessionConfig.R
\name{getConfig}
\alias{getConfig}
\title{Load configuration parameters}
\usage{
getConfig(.Object, ...)
}
\arguments{
\item{.Object}{a SessionConfig instance}

\item{file}{the file path to the configuration file, default is "~/.dataone/config.csv"}
}
\value{
The value of the requested parameter if found or NULL if not.
}
\description{
Load a configuration into the current configuration session.
}
\details{
This routine has the side effect of creating an R variable for each
paramater contained in the specified configuration file. If a configuration file
is not specified, then the default file is used. If configuration is being run for the first
time, then a default file is loaded from the dataone package.
}

