% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getD1Object}
\alias{getD1Object}
\title{Download a D1Object from the DataONE Federation}
\usage{
getD1Object(x, identifier, ...)
}
\arguments{
\item{x}{: D1Client}

\item{identifier}{: the identifier of the object to get}

\item{...}{(not yet used)}
}
\value{
the DataONE object
}
\description{
Download a D1Object from the DataONE Federation
}

