% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Object.R
\docType{methods}
\name{getData,D1Object,ANY-method}
\alias{getData,D1Object,ANY-method}
\title{Get the Contents of the Specified Data Object}
\usage{
\S4method{getData}{D1Object,ANY}(x, fileName = NA)
}
\arguments{
\item{x}{D1Object: the object from which to retrieve bytes of data}

\item{...}{(not yet used)}
}
\value{
character representation of the data
}
\description{
Get the Contents of the Specified Data Object
}
\author{
rnahf
}

