% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getData}
\alias{getData}
\title{Get the Contents of the Specified Data Object}
\usage{
getData(x, id, ...)
}
\arguments{
\item{x}{D1Object or DataPackage: the data structure from where to get the data}

\item{id}{Missing or character: if \code{"x"} is DataPackage, the identifier of the
package member to get data from}

\item{...}{(not yet used)}
}
\value{
character representation of the data
}
\description{
Get the Contents of the Specified Data Object
}
\author{
rnahf
}

