% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getDataObject}
\alias{getDataObject}
\title{Download a DataObject from the DataONE Federation}
\usage{
getDataObject(x, identifier, ...)
}
\arguments{
\item{x}{: D1Client}

\item{identifier}{: the identifier of the object to get}

\item{...}{(not yet used)}
}
\value{
the DataONE object
}
\description{
Download a DataObject from the DataONE Federation
}

