% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CNode.R
\name{getFormat}
\alias{getFormat}
\title{Get information for a single DataONE object format}
\usage{
getFormat(cnode, ...)
}
\arguments{
\item{cnode}{A CNode object instance}
}
\value{
Returns a dataframe of all object formats registered in the DataONE Object Format Vocabulary.
}
\description{
Get information for a single DataONE object format
}
\examples{
\dontrun{
cn <- CNode()
fmt <- getFormat(cn, "eml://ecoinformatics.org/eml-2.1.0")
cat(sprintf("format name: \%s\\n", fmt$name))
cat(sprintf("format type: \%s\\n", fmt$type))
cat(sprintf("format Id: \%s\\n", fmt$id))
}
}
\seealso{
http://mule1.dataone.org/ArchitectureDocs-current/apis/CN_APIs.html#CNCore.getFormat
}

