% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFormatId}
\alias{getFormatId}
\title{Get the FormatId of the D1Object}
\usage{
getFormatId(x, ...)
}
\arguments{
\item{x}{D1Object}

\item{...}{(not yet used)}
}
\value{
the formatId
}
\description{
Get the FormatId of the D1Object
}
\author{
rnahf
}

