% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getIdentifier}
\alias{getIdentifier}
\title{Get the Identifier of the D1Object}
\usage{
getIdentifier(x, ...)
}
\arguments{
\item{x}{D1Object}

\item{...}{(not yet used)}
}
\value{
the identifier
}
\description{
Get the Identifier of the D1Object
}
\author{
rnahf
}

