% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Client.R
\docType{methods}
\name{getMN}
\alias{getMN}
\alias{getMN,D1Client,ANY-method}
\alias{getMN,D1Client,character-method}
\title{Get a member node client based on its node identifier.}
\usage{
getMN(x, nodeid, ...)

\S4method{getMN}{D1Client,ANY}(x, nodeid, ...)

\S4method{getMN}{D1Client,character}(x, nodeid)
}
\arguments{
\item{x}{A D1Client object.}

\item{nodeid}{}

\item{...}{(Not yet used)}
}
\description{
Get a member node client based on its node identifier.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = D1Client,nodeid = ANY}: 

\item \code{x = D1Client,nodeid = character}: 
}}
\seealso{
\code{\link[=D1Client-class]{D1Client}}{ class description.}

\code{\link[=D1Client-class]{D1Client}}{ class description.}
}

