% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getMNodeId}
\alias{getMNodeId}
\title{Get for the member node identifier associated with this client.
One Member Node can be associated with the client as the default to which
data and metadata are written.}
\usage{
getMNodeId(x)
}
\description{
Get for the member node identifier associated with this client.
One Member Node can be associated with the client as the default to which
data and metadata are written.
}

