% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Node.R
\name{getQueryEngineDescription}
\alias{getQueryEngineDescription}
\title{Retrieve a description of a query engine from a coordinating node or member node}
\usage{
getQueryEngineDescription(node, queryEngineName)
}
\arguments{
\item{node}{The CNode or MNode to query}

\item{queryEngine}{The query engine name to get a description for.}
}
\value{
list The query engine description
}
\description{
Retrieve a description of a query engine from a coordinating node or member node
}

