% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CNode.R, R/D1Node.R
\docType{methods}
\name{getSystemMetadata,CNode,character-method}
\alias{getSystemMetadata}
\alias{getSystemMetadata,CNode,character-method}
\title{Get the metadata describing system properties associated with an object on a Coordinating Node.}
\usage{
\S4method{getSystemMetadata}{CNode,character}(node, pid)

getSystemMetadata(node, pid, ...)
}
\arguments{
\item{node}{The CNode instance from which the SystemMetadata will be downloaded}

\item{pid}{The identifier of the object}

\item{...}{(Not yet used.)}

\item{node}{The Node instance from which the SystemMetadata will be downloaded}

\item{pid}{The identifier of the object}
}
\value{
SystemMetadata for the object

SystemMetadata for the object
}
\description{
The SystemMetadata includes information about the identity, type, access control, and other system
level details about the object.

The SystemMetadata includes information about the identity, type, access control, and other system
level details about the object.
}
\details{
This operation acts as the 'public' anonymous user unless an X.509 certificate is
present in the default location of the file system, in which case the access will be authenticated.

This operation acts as the 'public' anonymous user unless an X.509 certificate is
present in the default location of the file system, in which case the access will be authenticated.
}
\section{Methods (by class)}{
\itemize{
\item \code{node = CNode,pid = character}: 
}}
\seealso{
\code{\link[=CNode-class]{CNode}}{ class description.}
}

