% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthenticationManager.R
\docType{methods}
\name{getToken}
\alias{getToken}
\alias{getToken,AuthenticationManager-method}
\title{Get the value of the DataONE Authentication Token, if one exists.}
\usage{
getToken(.Object)

\S4method{getToken}{AuthenticationManager}(.Object)
}
\arguments{
\item{.Object}{an AuthenticationManager instance}
}
\value{
The current authentication token.
}
\description{
Get the value of the DataONE Authentication Token, if one exists.
}

