% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CNode.R
\docType{methods}
\name{hasReservation}
\alias{hasReservation}
\alias{hasReservation,CNode-method}
\title{Get the list of nodes associated with a CN}
\usage{
hasReservation(cnode, ...)

\S4method{hasReservation}{CNode}(cnode, pid, subject = as.character(NA),
  quiet = TRUE)
}
\arguments{
\item{cnode}{The coordinating node to query for its registered Member Nodes}

\item{...}{Additional parameters.}

\item{pid}{The identifier that is being checked for existing as a reserved identifier or is in use as
an identifier for an existing object}

\item{subject}{The subject of the principal (user) that made the reservation. If not specified, then}

\item{quiet}{A logical value - if TRUE (the default) then informational messages are not printed.}
}
\value{
A logical value where TRUE means a reservation exists for the specified pid by the subject.
}
\description{
Get the list of nodes associated with a CN
}
\section{Methods (by class)}{
\itemize{
\item \code{CNode}: 
}}
\seealso{
\link{http://mule1.dataone.org/ArchitectureDocs-current/apis/CN_APIs.html#CNCore.hasReservation}
}

