% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Object.R
\docType{methods}
\name{initialize,D1Object-method}
\alias{initialize,D1Object-method}
\title{Initialize a D1Object}
\usage{
\S4method{initialize}{D1Object}(.Object, id = as.character(NA), data = NA,
  format = as.character(NA), mnNodeId = as.character(NA),
  filename = as.character(NA))
}
\arguments{
\item{.Object}{}

\item{id}{The identifier for the object}

\item{data}{An R object (data or metadata) that this D1Object contains.}

\item{format}{The Object format.}

\item{mnNodeId}{The DataONE node identifier associated with this object, i.e. "urn:node:KNB"}

\item{filename}{A filename that this D1Object contains.}
}
\description{
Initialize a D1Object
}

