% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SessionConfig.R
\docType{methods}
\name{initialize,SessionConfig-method}
\alias{initialize,SessionConfig-method}
\title{Initialize a SessionConfig object}
\usage{
\S4method{initialize}{SessionConfig}(.Object)
}
\description{
Initialize a SessionConfig object
}

