% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AuthenticationManager.R
\name{isAuthExpired}
\alias{isAuthExpired}
\title{CHeck if the currently valid authentication method has reached the expiratin time.}
\usage{
isAuthExpired(x, ...)
}
\arguments{
\item{x}{an Authentication instance}

\item{...}{(Not yet used.)}
}
\value{
A logical value: TRUE if authenentication has expired, FALSE if not.
}
\description{
CHeck if the currently valid authentication method has reached the expiratin time.
}

