% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CNode.R
\docType{methods}
\name{isAuthorized}
\alias{isAuthorized}
\alias{isAuthorized,CNode,character,character-method}
\title{Check if an action is authorized for the specified identifier}
\usage{
isAuthorized(cnode, id, action, ...)

\S4method{isAuthorized}{CNode,character,character}(cnode, id, action)
}
\arguments{
\item{cnode}{The node to send the request to.}

\item{id}{The DataONE identifier (pid or sid) to check access for.}

\item{action}{The DataONE action to check, possible values: "read", "write", "changePermission"}

\item{...}{(Not yet used)}
}
\value{
a logical, TRUE if the action is authorized, false if not.
}
\description{
Test if the user identified by the provided token has
authorization for operation on the specified object.
}
\details{
The identifer parameter may be either a DataONE persistant identifier (pid)
or series identifier (sid).
}
\section{Methods (by class)}{
\itemize{
\item \code{cnode = CNode,id = character,action = character}: 
}}
\examples{
\dontrun{
cn <- CNode("STAGING")
isAuthorized(cn, "doi:10.5072/FK2/LTER/sbclter.842.1", "read")
isAuthorized(cn, "doi:10.5072/FK2/LTER/sbclter.842.1", "write")
isAuthorized(cn, "doi:10.5072/FK2/LTER/sbclter.842.1", "changePermission")
}
}
\seealso{
\code{\link[=CNode-class]{CNode}}{ class description.}
}

