% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SessionConfig.R
\name{listConfig}
\alias{listConfig}
\title{List all currently defined configuration parameters}
\usage{
listConfig(.Object, ...)
}
\description{
All currently defined configuration parameter that were instatiated
with loadConfig() are listed
}

