% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CNode.R
\docType{methods}
\name{listNodes}
\alias{listNodes}
\alias{listNodes,CNode-method}
\title{Get the list of nodes associated with a CN}
\usage{
listNodes(cnode, ...)

\S4method{listNodes}{CNode}(cnode, url = as.character(NA), ...)
}
\arguments{
\item{cnode}{The coordinating node to query for its registered Member Nodes}

\item{...}{(Not yet used)}

\item{url}{Optianal - the url of the CN.}
}
\value{
the list of nodes in the DataONE CN environment
}
\description{
Get the list of nodes associated with a CN
}
\section{Methods (by class)}{
\itemize{
\item \code{CNode}: 
}}
\seealso{
\code{\link[=CNode-class]{CNode}}{ class description.}
}

