% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Node.R
\docType{methods}
\name{listObjects}
\alias{listObjects}
\alias{listObjects,D1Node-method}
\title{Retrieve the list of objects that match the search parameters}
\usage{
listObjects(node, ...)

\S4method{listObjects}{D1Node}(node, fromDate = as.character(NA),
  toDate = as.character(NA), formatId = as.character(NA),
  replicaStatus = as.logical(TRUE), start = as.integer(0),
  count = as.integer(1000))
}
\arguments{
\item{node}{The Node instance from which the SystemMetadata will be downloaded}

\item{...}{(Not yet used.)}

\item{fromDate}{Entries with a modified date greater than \code{'fromDate'} will be returned.
This value must be specified in ISO 8601 format, i.e. "YYYY-MM-DDTHH:MM:SS.mmm+00:00"}

\item{toDate}{Entries with a modified date less than \code{'toDate'} will be returned.
This value must be specified in ISO 8601 format, i.e. "YYYY-MM-DDTHH:MM:SS.mmm+00:00"}

\item{formatId}{The format to match, for example "eml://ecoinformatics.org/eml-2.1.1"}

\item{replicaStatus}{A logical value that determines if replica (object not on it's origin node) should be returned. Default is TRUE.}

\item{start}{An integer that specifies the first element of the result set that will be returned}

\item{count}{An integer that specifies how many results will be returned}
}
\value{
list Objects that met the search criteria

list Objects that met the search criteria
}
\description{
Retrieve the list of objects that match the search parameters
}
\details{
The list of objects that is returned is paged according to the \code{'start'} and
\code{'count'} values, so that large result sets can be returned over multiple calls.
}
\section{Methods (by class)}{
\itemize{
\item \code{D1Node}: 
}}
\seealso{
\url{http://mule1.dataone.org/ArchitectureDocs-current/apis/MN_APIs.html#MN_read.listObjects}
}

