% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Node.R
\docType{methods}
\name{listQueryEngines,D1Node-method}
\alias{listQueryEngines,D1Node-method}
\title{Query a node for the list of query engines available on the node}
\usage{
\S4method{listQueryEngines}{D1Node}(node)
}
\arguments{
\item{node}{The CNode or MNode instance to list the query engines for.}
}
\value{
list Objects that met the search criteria
}
\description{
Query a node for the list of query engines available on the node
}

