% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Node.R
\name{listQueryEngines}
\alias{listQueryEngines}
\title{List the query engines available for a DataONE member node or coordinating node}
\usage{
listQueryEngines(node, ...)
}
\arguments{
\item{node}{The CNode or MNode to list the query engines for.}
}
\value{
list The list of query engines.
}
\description{
List the query engines available for a DataONE member node or coordinating node
}

