% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MNode}
\alias{MNode}
\title{Create a MNode object representing a DataONE Member Node repository.}
\usage{
MNode(x)
}
\arguments{
\item{x}{a URI representing a node base URL, or a reference to a dataone::Node instance}
}
\value{
the MNode object
}
\description{
Construct an instance of MNode to provide mechanisms to access, create, and update data and
metadata objects on the associated Member Node.
}
\details{
If the \code{'x'} is a string, it is treated as a URI and an attempt to find an associated
Member Node at that base URL is attempted.  If \code{'x'} is a Node reference, then it is cast to a MNode
instance.  This typically is used from the getMNode() function from the CNode class, which is the preferred
way to retrieve an instance of an MNode.
}

