% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CertificateManager.R
\name{obscureCert}
\alias{obscureCert}
\title{Obscure the CILogon Client Certificate}
\usage{
obscureCert(x, ...)
}
\arguments{
\item{x}{a CertificateManager instance}
}
\value{
the modified CertificateManager instance
}
\description{
Obscures the x509 certificate that CILogon installs, effectively making future
interactions with the DataONE services public/anonymous.  This function simple renames an
existing certificate file to a known location, allowing 'public' operations.  Note, when the
client certificate is obscured via the renaming, you will not be able to create objects in
DataONE, or utilize any other methods that require authentication.
}
\seealso{
\code{\link{restoreCert}} is this method's inverse operation
}

