% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Node.R
\docType{methods}
\name{parseCapabilities}
\alias{parseCapabilities}
\alias{parseCapabilities,D1Node,XMLInternalElementNode-method}
\title{Construct a Node, using a passed in capabilities XML}
\usage{
parseCapabilities(node, xml, ...)

\S4method{parseCapabilities}{D1Node,XMLInternalElementNode}(node, xml)
}
\arguments{
\item{node}{The node to which capabilities should be applied.}

\item{xml}{The XML capabilities representing the node to be created}

\item{...}{(not yet used)}
}
\value{
The Node object with modified capabilities properties from the XML
}
\description{
Construct a Node, using a passed in capabilities XML
}

