% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Node.R
\name{parseSolrResult}
\alias{parseSolrResult}
\title{Parse Solr output into an R list}
\usage{
parseSolrResult(doc, ...)
}
\arguments{
\item{result}{The Solr result to parse, in XML format}
}
\value{
resultList The Solr result as an R list
}
\description{
Solr output that is specified with a writer type of XML \code{'&wt="xml"'}
}

