% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Node.R
\docType{methods}
\name{ping,D1Node-method}
\alias{ping,D1Node-method}
\title{Test if a node is online and accepting DataONE requests}
\usage{
\S4method{ping}{D1Node}(node)
}
\arguments{
\item{node}{The CNode or MNode to check.}
}
\value{
list Objects that met the search criteria
}
\description{
Test if a node is online and accepting DataONE requests
}

