% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CNode.R
\name{query}
\alias{query}
\title{Search DataONE for data and metadata objects}
\usage{
query(cnode, solrQuery, ...)
}
\arguments{
\item{cnode}{The coordinating node to query}

\item{solrQuery}{The query parameters to be searched, either as a URL-encoded string or as list with named attributes}
}
\value{
search results
}
\description{
Search DataONE for data and metadata objects
}
\details{
Use SOLR syntax to search the DataONE federation of data repositories for matching data.
Any lucene reserved characters must be escaped with backslash. If solrQuery is a list,
it is expected to have field names as attributes and search values as the values in the list.
}

