% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CNode.R
\docType{methods}
\name{resolve}
\alias{resolve}
\alias{resolve,CNode,character-method}
\title{Get a list of coordinating nodes holding a given pid}
\usage{
resolve(cnode, pid)

\S4method{resolve}{CNode,character}(cnode, pid)
}
\arguments{
\item{cnode}{a valid CNode object}

\item{pid}{the id of the identified object}
}
\description{
Returns a list of nodes (MNs or CNs) known to hold copies of the object identified by id.
}
\examples{
\dontrun{
cn <- CNode("SANDBOX")
id <- "doi:10.5072/FK2/LTER/knb-lter-gce.100.15"
resolve(cn,id)
}
}
\author{
hart
}

