% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AuthenticationManager.R
\docType{methods}
\name{restoreAuth}
\alias{restoreAuth}
\alias{restoreAuth,AuthenticationManager-method}
\title{Get DataONE Identity as Stored in the CILogon Certificate.}
\usage{
restoreAuth(x, ...)

\S4method{restoreAuth}{AuthenticationManager}(x)
}
\arguments{
\item{x}{an Authentication instance}
}
\value{
The expiration date for the current authentication mechanism being used.
}
\description{
Get DataONE Identity as Stored in the CILogon Certificate.
}
\section{Methods (by class)}{
\itemize{
\item \code{AuthenticationManager}: 
}}

