% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CertificateManager.R
\name{restoreCert}
\alias{restoreCert}
\title{Restore the CILogon client certificate by renaming it to its original location}
\usage{
restoreCert(x, ...)
}
\arguments{
\item{x}{a CertificateManager instance}

\item{...}{(Not yet used)}
}
\value{
the modified CertificateManager instance
}
\description{
Restores the x509 certificate that CILogon installs, which allows future
interactions with nodes to be authenticated with the certificate.  This function simply
renames an obscured certificate file to its original location, allowing authenticated
operations.
}
\seealso{
\code{\link{obscureCert}} is this method's inverse operation
}

