% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{serialize}
\alias{serialize}
\alias{serialize,SystemMetadata-method}
\title{serialize metadata}
\usage{
serialize(sysmeta, ...)

\S4method{serialize}{SystemMetadata}(sysmeta)
}
\arguments{
\item{sysmeta}{the SystemMetadata instance to be serialized}
}
\value{
the character string representing a SystemMetadata object
}
\description{
Serialize an XML representation of system metadata.
}
\author{
jones
}

