% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SessionConfig.R
\docType{class}
\name{SessionConfig-class}
\alias{SessionConfig-class}
\alias{getConfig,SessionConfig-method}
\alias{listConfig,SessionConfig-method}
\alias{loadConfig,SessionConfig-method}
\alias{saveConfig,SessionConfig-method}
\alias{setConfig,SessionConfig,character-method}
\alias{unloadConfig,SessionConfig-method}
\title{SessionConfig provides methods to get, set, load, unload dataone configuration parameters.}
\usage{
\S4method{loadConfig}{SessionConfig}(.Object, file = as.character(NA))

\S4method{unloadConfig}{SessionConfig}(.Object)

\S4method{saveConfig}{SessionConfig}(.Object,
  file = .Object@DefaultConfigFilepath)

\S4method{setConfig}{SessionConfig,character}(.Object,
  name = as.character(NA), value)

\S4method{getConfig}{SessionConfig}(.Object, name = as.character(NA))

\S4method{listConfig}{SessionConfig}(.Object, name = as.character(NA))
}
\description{
SessionConfig provides methods to get, set, load, unload dataone configuration parameters.
}
\section{Methods (by generic)}{
\itemize{
\item \code{loadConfig}: 

\item \code{unloadConfig}: 

\item \code{saveConfig}: 

\item \code{setConfig}: 

\item \code{getConfig}: 

\item \code{listConfig}: 
}}
\section{Slots}{

\describe{
\item{\code{DefaultConfigFilepath}}{the default config file to load if no other filename is specified}

\item{\code{InitialConfigFilepath}}{the initial config file if a config file has never been created}
}}

