% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SessionConfig.R
\name{setConfig}
\alias{setConfig}
\title{Load configuration parameters}
\usage{
setConfig(.Object, name, value)
}
\arguments{
\item{.Object}{a SessionConfig instance}

\item{file}{the file path to the configuration file, default is "~/.dataone/config.csv"}
}
\description{
Load a configuration into the current configuration session.
}
\details{
This routine has the side effect of creating an R variable for each
paramater contained in the specified configuration file. If a configuration file
is not specified, then the default file is used. If Configuration is being run for the first
time, then a default file is loaded from the dataone package.
}

