% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CNode.R
\docType{methods}
\name{setObsoletedBy}
\alias{setObsoletedBy}
\alias{setObsoletedBy,CNode,character-method}
\title{Set a pid as being obsoleted by another pid}
\usage{
setObsoletedBy(cnode, pid, obsoletedByPid, ...)

\S4method{setObsoletedBy}{CNode,character}(cnode, pid, obsoletedByPid,
  serialVersion)
}
\arguments{
\item{pid}{The identifier of the object to be obsoleted}

\item{obsoletedByPid}{The identifier of the object that obsoletes the object identified by pid.}

\item{...}{(Not yet used)}

\item{serialVersion}{The serial version of the system metadata of the pid being obsoleted.}

\item{node}{The CNode instance on which the object will be created}
}
\value{
TRUE if the pid was obsoleted, otherwise FALSE is returned
}
\description{
Updates the SystemMetadata 'obsoletedBy' property for an object, indicating that the object
specified by pid has been obsoleted by the identifier in obsoletedByPid.
CILogon \url{https://cilogon.org/?skin=DataONE}.  See \code{\link{{CertificateManager}}} for details.
}
\section{Methods (by class)}{
\itemize{
\item \code{cnode = CNode,pid = character}: 
}}
\seealso{
\code{\link[=CNode-class]{CNode}}{ class description.}
}

