% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CertificateManager.R
\name{showClientSubject}
\alias{showClientSubject}
\title{Get DataONE Identity as Stored in the CILogon Certificate.}
\usage{
showClientSubject(x, ...)
}
\arguments{
\item{x}{a CertificateManager instance}
}
\value{
the DataONE Subject that is your client's identity
}
\description{
Returns Your Identity according to DataONE (and CILogon) as provided in the Subject
field of the X.509 certificate.  The value is a Distinguished Name, and can be used in all fields that
require a user identity for access control authorization. If the certificate is missing on expired, then
the subject 'public' is returned.
}

