% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SessionConfig.R
\name{unloadConfig}
\alias{unloadConfig}
\title{Unload configuration parameters}
\usage{
unloadConfig(.Object, ...)
}
\arguments{
\item{.Object}{a Configuration instance}

\item{file}{the file path to the configuration file, default is "~/.dataone/config.csv"}
}
\description{
Unload a configuration
}
\details{
The R environment that contained the configuration parameters is destroyed
}

