% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/D1Client.R
\name{upload}
\alias{upload}
\title{Upload an object to the DataONE System.}
\usage{
upload(x, ...)
}
\arguments{
\item{x}{: D1Client to do the uploading}

\item{...}{(not yet used)}

\item{object}{: the object to create in DataONE}
}
\value{
identifier of the uploaded object if success, otherwise FALSE
}
\description{
Uploads a DataObject on the MemberNode determined by the object's systemMetadata.
Values in the object's SystemMetadata are used to determine where the object is
is uploaded, its identifier, format, owner, access policies, and other relevant
metadata about the object.
}

