% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Client.R
\name{uploadDataObject}
\alias{uploadDataObject}
\title{Upload a DataObject to a DataONE member node.}
\usage{
uploadDataObject(x, do, ...)
}
\arguments{
\item{x}{A D1Client instance.}

\item{do}{The DataObject instance to be uploaded to DataONE.}

\item{replicate}{A value of type \code{"logical"}, if TRUE then DataONE will replicate this object to other member nodes}

\item{numberReplicas}{A value of type \code{"numeric"}, for number of supported replicas.}

\item{preferredNodes}{A list of \code{"character"}, each of which is the node identifier for a node to which a replica should be sent.}

\item{accessRules}{Access rules of \code{'data.frame'} that will be added to the access policy}
}
\value{
id The id of the DataObject that was uploaded
}
\description{
Upload a DataObject to a DataONE member node.
}
\seealso{
\code{\link[=D1Client-class]{D1Client}}{ class description.}
}

