% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MNode.R
\name{uploadDataPackage}
\alias{uploadDataPackage}
\title{Upload a DataPackage to a DataONE member node}
\usage{
uploadDataPackage(mn, dp, ...)
}
\arguments{
\item{mn}{An MNode object that represents the member node to upload the package to.}

\item{dataPackage}{The DataPackage instance to be submitted to DataONE for creation.}

\item{replicate}{A value of type \code{"logical"}, if TRUE then DataONE will replicate this object to other member nodes}

\item{numberReplicas}{A value of type \code{"numeric"}, for number of supported replicas.}

\item{preferredNodes}{A list of \code{"character"}, each of which is the node identifier for a node to which a replica should be sent.}

\item{public}{A \code{'logical'}, if TRUE then all objects in this package will be accessible by any user}

\item{accessRules}{Access rules of \code{'data.frame'} that will be added to the access policy}
}
\value{
id The identifier of the resource map for this data package
}
\description{
Upload all DataObjects contained in the DataPackage by calling uploadDataObject()
on each of the members. Also a resourceMap object is created from the
recorded relationships between DataObjects, and this is uploaded as well.
}
\details{
The DataPackage describes the collection of data object and their associated
metadata object, with the relationships and members serialized into a document
stored under, and retrievable with, the packageId as it's own distinct object.

Members are created serially, and most errors in creating one object will
interrupt the create process for the whole, with the result that some members will
be created, and the remainder not.
}

