% Generated by roxygen2 (4.0.2): do not edit by hand
\name{archive}
\alias{archive}
\title{Archive an object on a Member Node, which hides it from casual searches.}
\usage{
archive(mnode, pid, ...)
}
\arguments{
\item{pid}{The identifier of the object to be created}

\item{node}{The MNode instance on which the object will be created}
}
\value{
XML describing the result of the operation, inlcuding the identifier if successful
}
\description{
This method provides the ability to archive a data or metadata object on the Member Node
provided in the \code{'mnode'} parameter.  Archiving removes the object from DataONE search functions,
thereby making it more difficult to find without completely removing the object.  Archive is intended
for objects that should not be used by current researchers, but for which there is a desire to maintain
a historical record, such as when journal articles might cite the object.  Users can still obtain the
contents of archived objects if they have the identifier, but will not discover it through searches.
}
\details{
This operation requires an X.509 certificate to be present in the default location of the file
system. This certificate provides authentication credentials from
CILogon \url{https://cilogon.org/?skin=DataONE}.  See \code{\link{{CertificateManager}}} for details.
}
\seealso{
\url{http://mule1.dataone.org/ArchitectureDocs-current/apis/MN_APIs.html#MNStorage.archive}
}

