% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Object.R
\docType{methods}
\name{asDataFrame}
\alias{asDataFrame}
\alias{asDataFrame,D1Object,ANY-method}
\alias{asDataFrame,D1Object,AbstractTableDescriber-method}
\alias{asDataFrame,D1Object,D1Object-method}
\title{return the D1Object data as a data.frame.}
\usage{
asDataFrame(x, reference, ...)

\S4method{asDataFrame}{D1Object,D1Object}(x, reference, ...)

\S4method{asDataFrame}{D1Object,AbstractTableDescriber}(x, reference, ...)

\S4method{asDataFrame}{D1Object,ANY}(x, reference, ...)
}
\arguments{
\item{x}{A D1Object}

\item{reference}{A reference to a D1Object}

\item{...}{(Additional parameters)}
}
\description{
This method uses the provided metadata reference object for instructions on
how to parse the data table (which parameters to set)
'reference' is the metadata D1Object that gives instruction on how to read the data
into the dataFrame
}

