% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CNode.R
\docType{methods}
\name{get,CNode,character-method}
\alias{describe,CNode,character-method}
\alias{get,CNode,character-method}
\alias{getSystemMetadata,CNode,character-method}
\title{Get the bytes associated with an object on this Coordinating Node.}
\usage{
\S4method{get}{CNode,character}(node, pid)

\S4method{getSystemMetadata}{CNode,character}(node, pid)

\S4method{describe}{CNode,character}(node, pid)
}
\arguments{
\item{node}{The CNode instance from which the pid will be downloaded}

\item{pid}{The identifier of the object to be downloaded}

\item{node}{The CNode instance from which the SystemMetadata will be downloaded}

\item{pid}{The identifier of the object}

\item{mnode}{The CNode instance from which the identifier will be generated}

\item{pid}{Identifier for the object in question. May be either a PID or a SID. Transmitted as
part of the URL path and must be escaped accordingly.}
}
\value{
the bytes of the object

SystemMetadata for the object

A list of header elements
}
\description{
Get the bytes associated with an object on this Coordinating Node.

The SystemMetadata includes information about the identity, type, access control, and other system
level details about the object.

This method provides a lighter weight mechanism than getSystemMetadata() for a client to
determine basic properties of the referenced object.
}
\details{
This operation acts as the 'public' anonymous user unless an X.509 certificate is
present in the default location of the file system, in which case the access will be authenticated.

This operation acts as the 'public' anonymous user unless an X.509 certificate is
present in the default location of the file system, in which case the access will be authenticated.
}
\section{Methods (by generic)}{
\itemize{
\item \code{get}: 

\item \code{getSystemMetadata}: 

\item \code{describe}: 
}}
\author{
Scott Chamberlain
}
\seealso{
\url{http://mule1.dataone.org/ArchitectureDocs-current/apis/CN_APIs.html#CNRead.get}

\url{http://mule1.dataone.org/ArchitectureDocs-current/apis/CN_APIs.html#CNRead.getSystemMetadata}

http://mule1.dataone.org/ArchitectureDocs-current/apis/CN_APIs.html#CNRead.describe
}

