% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CNode.R
\name{CNode}
\alias{CNode}
\title{Create a CNode object}
\usage{
CNode(env, ...)
}
\arguments{
\item{env}{The label for the DataONE environment to be using ('PROD','STAGING','SANDBOX','DEV')}

\item{...}{(not yet used)}
}
\value{
the CNode object representing the DataONE environment
}
\description{
Create a CNode object
}
\author{
jones
}

