% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNode.R
\docType{methods}
\name{CNode}
\alias{CNode}
\alias{CNode,ANY-method}
\alias{CNode,character-method}
\title{Create a CNode object.}
\usage{
CNode(env, ...)

\S4method{CNode}{ANY}()

\S4method{CNode}{character}(env)
}
\arguments{
\item{env}{The label for the DataONE environment to be using ('PROD','STAGING', 'STAGING2,'SANDBOX', 'SANDBOX2','DEV', 'DEV2')}

\item{...}{(not yet used)}
}
\value{
the CNode object representing the DataONE environment
}
\description{
Create a CNode object.
}
\examples{
\dontrun{
cn <- CNode("PROD")
}
}
\seealso{
\code{\link[=CNode-class]{CNode}}{ class description.}
}

