% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/D1Client.R
\name{convert.csv}
\alias{convert.csv}
\title{Convert a DataFrame to Standard CSV}
\usage{
convert.csv(x, ...)
}
\arguments{
\item{...}{additional params passed to write.csv}

\item{df}{the dataFrame}
}
\value{
the dataframe serialized as a .csv
}
\description{
Convert a DataFrame to Standard CSV
}

