% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createD1Object}
\alias{createD1Object}
\title{Create the Object in the DataONE System}
\usage{
createD1Object(x, d1Object, ...)
}
\arguments{
\item{x}{: D1Client}

\item{d1Object}{: the object to create in DataONE}

\item{...}{(not yet used)}
}
\value{
TRUE if success
}
\description{
Creates a D1Object on the MemberNode determined by the object's systemMetadata.
}

