% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Client.R
\name{createDataPackage}
\alias{createDataPackage}
\title{Create a DataPackage on a DataONE Member Node}
\usage{
createDataPackage(x, dataPackage, ...)
}
\arguments{
\item{x}{A D1Client instance.}

\item{dataPackage}{The DataPackage instance to be submitted to DataONE for creation.}

\item{...}{(not yet used)}
}
\description{
This method has been superceded by \link{updateDataPacakge}.
}

