% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Client.R
\docType{methods}
\name{createDataPackage}
\alias{createDataPackage}
\alias{createDataPackage,D1Client,DataPackage-method}
\title{Create a DataPackage on a DataONE Member Node}
\usage{
createDataPackage(x, dataPackage, ...)

\S4method{createDataPackage}{D1Client,DataPackage}(x, dataPackage)
}
\arguments{
\item{x}{A D1Client instance.}

\item{dataPackage}{The DataPackage instance to be submitted to DataONE for creation.}

\item{...}{(not yet used)}
}
\description{
This method has been superceded by \code{\link{uploadDataPackage}}
}
\seealso{
\code{\link[=D1Client-class]{D1Client}}{ class description.}
}

