% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{D1Client,ANY,ANY-method}
\alias{D1Client,ANY,ANY-method}
\title{Construct a D1Client, using default env ("PROD") and nodeid ("")}
\usage{
\S4method{D1Client}{ANY,ANY}()
}
\value{
the D1Client object representing the DataONE environment
}
\description{
Construct a D1Client, using default env ("PROD") and nodeid ("")
}

