% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Client.R
\docType{class}
\name{D1Client-class}
\alias{D1Client,ANY,ANY-method}
\alias{D1Client,character,ANY-method}
\alias{D1Client,character,character-method}
\alias{D1Client-class}
\title{The D1Client class contains methods that perform high level DataONE tasks}
\usage{
\S4method{D1Client}{ANY,ANY}()

\S4method{D1Client}{character,ANY}(env, mNodeid, ...)

\S4method{D1Client}{character,character}(env, mNodeid)
}
\description{
The methods in the D1Client class call the low level DataONE API to
perform involved tasks such as uploading all the packages in a DataPackage (i.e
\code{\link{uploadDataPackage}})
}
\section{Methods (by generic)}{
\itemize{
\item \code{D1Client}: 

\item \code{D1Client}: 

\item \code{D1Client}: 
}}
\section{Slots}{

\describe{
\item{\code{endpoint}}{The baseurl of the CN in that en  vironment}

\item{\code{mn.nodeid}}{The NodeReference for the 'home' MemberNode for this application, where creates/updates will happen.}

\item{\code{client}}{The reference to the internally held Java D1Client instance}

\item{\code{session}}{The reference to the internally held Java Session instance}
}}
\section{Methods}{

\itemize{
 \item{\code{\link{D1Client}}}{: Construct a D1Client object.}
 \item{\code{\link{getPackage}}}{: Download a data package from the DataONE Federation.}
 \item{\code{\link{getD1Object}}}{: Download a data object from the DataONE Federation.}
 \item{\code{\link{getDataObject}}}{: Get the data content of a specified data object}
 \item{\code{\link{d1SolrQuery}}}{: A method to query the DataONE solr endpoint of the Coordinating Node.}
 \item{\code{\link{d1IdentifierSearch}}}{: Query the DataONE Solr endpoint of the Coordinating Node.}
 \item{\code{\link{reserveIdentifier}}}{: Reserve a unique identifier in the DataONE Network.}
 \item{\code{\link{createDataPackage}}}{: Create a DataPackage on a DataONE Member Node}
 \item{\code{\link{getEndpoint}}}{: Return the URL endpoint for the DataONE Coordinating Node}
 \item{\code{\link{getMNodeId}}}{: Get the member node identifier associated with this D1Client object.}
 \item{\code{\link{getMN}}}{: Get a member node client based on its node identifier.}
 \item{\code{\link{uploadDataPackage}}}{: Upload a DataPackage to a DataONE member node.}
 \item{\code{\link{uploadDataObject}}}{: Upload a DataObject to a DataONE member node..}
 \item{\code{\link{listMemberNodes}}}{: List DataONE Member Nodes.}
 \item{\code{\link{convert.csv}}}{: Convert a DataFrame to Standard CSV.}
 \item{\code{\link{encodeUrlQuery}}}{: Encode the Input for a URL Query Segment.}
 \item{\code{\link{encodeUrlPath}}}{: Encode the Input for a URL Path Segment.}
}
}
\seealso{
\code{\link{dataone}}{ package description.}

\code{\link[=D1Client-class]{D1Client}}{ class description.}

\code{\link[=D1Client-class]{D1Client}}{ class description.}

\code{\link[=D1Client-class]{D1Client}}{ class description.}
}

