% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Client.R
\docType{methods}
\name{initialize,D1Client-method}
\alias{D1Client-initialize}
\alias{initialize,D1Client-method}
\title{Initialize a D1Client object}
\usage{
\S4method{initialize}{D1Client}(.Object, cn = NA, mn = NA,
  env = as.character(NA), mNodeid = as.character(NA))
}
\description{
Initialize a D1Client object
}
\seealso{
\code{\link[=D1Client-class]{dataone}}{ class description.}
}

