% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Client.R
\name{D1Client}
\alias{D1Client}
\title{The DataONE client class used to downlaod, update and search for data in the DataONE network.}
\usage{
D1Client(env, mNodeid, ...)
}
\arguments{
\item{env}{The label for the DataONE environment to be using ('PROD','STAGING','SANDBOX','DEV')}

\item{mNodeid}{The node Id of the application's 'home' node.  Should be already registered to the corresponding 'env'}

\item{...}{(not yet used)}
}
\value{
the D1Client object representing the DataONE environment
}
\description{
The DataONE client class used to downlaod, update and search for data in the DataONE network.
}
\seealso{
\code{\link[=D1Client-class]{D1Client}}{ class description.}
}

