% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Client.R
\docType{methods}
\name{d1IdentifierSearch}
\alias{d1IdentifierSearch}
\alias{d1IdentifierSearch,D1Client,character-method}
\title{Query the DataONE Solr endpoint of the Coordinating Node}
\usage{
d1IdentifierSearch(x, solrQuery)

\S4method{d1IdentifierSearch}{D1Client,character}(x, solrQuery)
}
\arguments{
\item{x}{D1Client: representing the DataONE environment being queried}

\item{solrQuery}{character: a query string}
}
\value{
a vector of identifiers found
}
\description{
The DataONE CN Solr query engine is searched using
the provided query string.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = D1Client,solrQuery = character}: 
}}
\examples{
\dontrun{
client <- new("D1Client")
result <- d1IdentifierSearch(client,q="species population diversity")
}
}
\seealso{
\code{\link[=D1Client-class]{D1Client}}{ class description.}
}

