% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{D1Object-class}
\alias{D1Object-class}
\title{D1Object (Debprecated) is a representation of a DataObject.}
\description{
D1Object has been deprecated in favor of datapackage::DataObject, which provides
a wrapper for data and associated SystemMetadata.
}
\section{Slots}{

\describe{
\item{\code{proxyObject}}{A backing instance of a DataObject, to which all methods and state are proxied}
}}
\examples{
\dontrun{
cn <- CNode("STAGING2")
mn <- getMNode(cn, "urn:node:mnTestKNB")
}
}
\author{
Matthew Jones
}
\keyword{classes}

