% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/D1Node.R
\name{describe}
\alias{describe}
\title{This method provides a lighter weight mechanism than getSystemMetadata() for a client to
determine basic properties of the referenced object.}
\usage{
describe(node, pid, ...)
}
\arguments{
\item{pid}{Identifier for the object in question. May be either a PID or a SID. Transmitted as
part of the URL path and must be escaped accordingly.}

\item{mnode}{The Node instance from which the identifier will be generated}
}
\value{
A list of header elements
}
\description{
This method provides a lighter weight mechanism than getSystemMetadata() for a client to
determine basic properties of the referenced object.
}
\author{
Scott Chamberlain
}

