% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CNode.R, R/D1Node.R, R/MNode.R
\docType{methods}
\name{describe,CNode,character-method}
\alias{describe}
\alias{describe,CNode,character-method}
\alias{describe,MNode,character-method}
\title{Efficiently get systemmetadat for an object.}
\usage{
\S4method{describe}{CNode,character}(node, pid)

describe(node, pid, ...)

\S4method{describe}{MNode,character}(node, pid)
}
\arguments{
\item{node}{The MNode instance from which the identifier will be generated}

\item{pid}{Identifier for the object in question. May be either a PID or a SID. Transmitted as
part of the URL path and must be escaped accordingly.}

\item{...}{(Not yet used)}
}
\value{
A list of header elements
}
\description{
This method provides a lighter weight mechanism than getSystemMetadata() for a client to
determine basic properties of the referenced object.
}
\section{Methods (by class)}{
\itemize{
\item \code{node = CNode,pid = character}: 

\item \code{node = MNode,pid = character}: 
}}
\examples{
\dontrun{
mn_uri <- "https://knb.ecoinformatics.org/knb/d1/mn/v1"
mn <- MNode(mn_uri)
pid <- "knb.473.1"
describe(mn, pid)
describe(mn, "adfadf") # warning message when wrong pid
}
}
\seealso{
\url{http://mule1.dataone.org/ArchitectureDocs-current/apis/MN_APIs.html#MNRead.describe}
}

