% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Node.R
\name{encodeSolr}
\alias{encodeSolr}
\title{Encode the input for Solr Queries}
\usage{
encodeSolr(segment, ...)
}
\arguments{
\item{segment}{: a string to encode}
}
\value{
the encoded form of the input
}
\description{
Treating all special characters and spaces as literals, backslash escape special
characters, and double-quote if necessary.
}
\examples{
encodeSolr("this & that")
}

