% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Client.R
\docType{methods}
\name{encodeUrlPath}
\alias{encodeUrlPath}
\alias{encodeUrlPath,D1Client,character-method}
\title{Encode the Input for a URL Path Segment.}
\usage{
encodeUrlPath(x, pathSegment, ...)

\S4method{encodeUrlPath}{D1Client,character}(x, pathSegment, ...)
}
\arguments{
\item{x}{A D1Client object}

\item{pathSegment}{: a string to encode}

\item{...}{(Not yet used.)}
}
\value{
the encoded form of the input
}
\description{
Encodes the characters of the input so they are not interpretted as reserved
characters in url strings.  Will also encode non-ASCII unicode characters.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = D1Client,pathSegment = character}: 
}}
\examples{
\dontrun{
fullyEncodedPath <- paste0("cn/v1/object/",
    encodeUrlPath("doi:10.6085/AA/YBHX00_XXXITBDXMMR01_20040720.50.5"))
}
}
\seealso{
\code{\link[=D1Client-class]{D1Client}}{ class description.}
}

