% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Node.R
\name{get}
\alias{get}
\title{Get the bytes associated with an object on this Node.}
\usage{
get(node, pid, ...)
}
\arguments{
\item{node}{The Node instance from which the pid will be downloaded}

\item{pid}{The identifier of the object to be downloaded}
}
\value{
the bytes of the object
}
\description{
Get the bytes associated with an object on this Node.
}
\details{
This operation acts as the 'public' anonymous user unless an X.509 certificate is
present in the default location of the file system, in which case the access will be authenticated.
}

