% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AuthenticationManager.R
\docType{methods}
\name{getAuthExpires}
\alias{getAuthExpires}
\alias{getAuthExpires,AuthenticationManager-method}
\title{Get DataONE Identity as Stored in the CILogon Certificate.}
\usage{
getAuthExpires(x, ...)

\S4method{getAuthExpires}{AuthenticationManager}(x)
}
\arguments{
\item{x}{an AuthenticationManager instance}

\item{...}{(Not yet used.)}
}
\value{
The expiration date for the current authentication mechanism being used.
}
\description{
Get DataONE Identity as Stored in the CILogon Certificate.
}
\section{Methods (by class)}{
\itemize{
\item \code{AuthenticationManager}: 
}}

