% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AuthenticationManager.R
\name{getAuthExpires}
\alias{getAuthExpires}
\title{Get DataONE Identity as Stored in the CILogon Certificate.}
\usage{
getAuthExpires(x, ...)
}
\arguments{
\item{x}{an Authentication instance}
}
\value{
The expiration date for the current authentication mechanism being used.
}
\description{
Get DataONE Identity as Stored in the CILogon Certificate.
}

