% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AuthenticationManager.R
\docType{methods}
\name{getAuthToken}
\alias{getAuthTo}
\alias{getAuthToken}
\alias{getAuthToken,AuthenticationManager-method}
\title{Get the value of the DataONE Authentication Token, if one exists.}
\usage{
getAuthToken(x, ...)

\S4method{getAuthToken}{AuthenticationManager}(x)
}
\arguments{
\item{x}{an AuthenticationManager instance}
}
\value{
The current authentication token.
}
\description{
Get the value of the DataONE Authentication Token, if one exists.
}
\section{Methods (by class)}{
\itemize{
\item \code{AuthenticationManager}: 
}}

