% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNode.R
\docType{methods}
\name{getCapabilities}
\alias{getCapabilities}
\alias{getCapabilities,MNode-method}
\title{Get the node capabilities description, and store the information in the MNode.}
\usage{
getCapabilities(mnode, ...)

\S4method{getCapabilities}{MNode}(mnode)
}
\arguments{
\item{mnode}{The node identifier with which this node is registered in DataONE}

\item{...}{(Not yet used.)}
}
\value{
the MNode object representing the DataONE environment
}
\description{
Access the DataONE getCapabilities() service for the Member Node, which returns an XML
decription of the repository and the services it offers.
}
\seealso{
\url{http://mule1.dataone.org/ArchitectureDocs-current/apis/MN_APIs.html#MN_core.getCapabilities}
}

