% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CNode.R, R/D1Node.R, R/MNode.R
\docType{methods}
\name{getChecksum,CNode,character-method}
\alias{getChecksum}
\alias{getChecksum,CNode,character-method}
\alias{getChecksum,MNode,character-method}
\title{Get the checksum for the data object associated with the specified pid.}
\usage{
\S4method{getChecksum}{CNode,character}(node, pid, ...)

getChecksum(node, pid, ...)

\S4method{getChecksum}{MNode,character}(node, pid,
  checksumAlgorithm = "SHA-1")
}
\arguments{
\item{node}{The CNode instance from which the checksum will be retrieved}

\item{pid}{The identifier of the object}

\item{...}{(Not yet used)}

\item{checksumAlgorithm}{The algorithm used to calculate the checksum. Default="SHA-1"}
}
\value{
character the checksum value, with the checksum algorithm as the attribute "algorithm"
}
\description{
A checksum is calculated for an object when it is uploaded to DataONE and
is submitted with the object's system metadata. The \code{'getChecksum'} method retrieves
the checksum from the specified coordinating node
}
\section{Methods (by class)}{
\itemize{
\item \code{node = CNode,pid = character}: 

\item \code{node = MNode,pid = character}: 
}}
\seealso{
\code{\link[=D1Node-class]{D1Node}}{ class description.}
}

