% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Client.R
\docType{methods}
\name{getD1Object}
\alias{getD1Object}
\alias{getD1Object,D1Client-method}
\title{Download a D1Object from the DataONE Cloud}
\usage{
getD1Object(x, identifier, ...)

\S4method{getD1Object}{D1Client}(x, identifier)
}
\arguments{
\item{x}{A D1Client instance}

\item{identifier}{The identifier of the object to download from DatONE}

\item{...}{(not yet used)}
}
\value{
A DataONE object
}
\description{
An objectd is download from the DataONE network for the identifier that is provided.
}
\section{Methods (by class)}{
\itemize{
\item \code{D1Client}: 
}}
\seealso{
\code{\link[=D1Client-class]{D1Client}}{ class description.}
}

