% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Object.R
\docType{methods}
\name{getData,D1Object,ANY-method}
\alias{getData,D1Object,ANY-method}
\title{Get the data content of a specified data object}
\usage{
\S4method{getData}{D1Object,ANY}(x, id)
}
\arguments{
\item{x}{D1Object the data structure from where to get the data}

\item{id}{Missing or character: if \code{'x'} is DataPackage, the identifier of the package member to get data from}
}
\description{
Get the data content of a specified data object
}
\section{Methods (by class)}{
\itemize{
\item \code{x = D1Object,id = ANY}: 
}}

