% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getDataObject}
\alias{getDataObject}
\title{Download a DataObject from the DataONE Federation.}
\usage{
getDataObject(x, identifier, ...)
}
\arguments{
\item{x}{: D1Client}

\item{identifier}{: the identifier of the object to get}

\item{...}{(not yet used)}
}
\value{
the DataONE object
}
\description{
A convenience method to download a data object and its associated SystemMetadata, wrapped
in a DataObject class.
}
\details{
This method performs multiple underlying calls to the DataONE repository network.
CN.resolve() is called to locate the object on one or more repositories, and then each of these
is accessed until success at downloading the associated SystemMetadata and data bytes, which are
finally wrapped in a DataObject and returned. Replaces previous getD1Object() method in the version 1
dataone library.
}

