% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CNode.R
\docType{methods}
\name{getFormat}
\alias{getFormat}
\alias{getFormat,CNode-method}
\title{Get information for a single DataONE object format}
\usage{
getFormat(cnode, ...)

\S4method{getFormat}{CNode}(cnode, formatId)
}
\arguments{
\item{cnode}{A CNode object instance}
}
\value{
A dataframe of all object formats registered in the DataONE Object Format Vocabulary.
}
\description{
Get information for a single DataONE object format
}
\section{Methods (by class)}{
\itemize{
\item \code{CNode}: 
}}
\examples{
\dontrun{
cn <- CNode()
fmt <- getFormat(cn, "eml://ecoinformatics.org/eml-2.1.0")
cat(sprintf("format name: \%s\\n", fmt$name))
cat(sprintf("format type: \%s\\n", fmt$type))
cat(sprintf("format Id: \%s\\n", fmt$id))
}
}
\seealso{
\code{\link[=CNode-class]{CNode}}{ class description.}
}

