% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Object.R
\docType{methods}
\name{getFormatId,D1Object-method}
\alias{getFormatId,D1Object-method}
\title{Get the FormatId of the DataObject}
\usage{
\S4method{getFormatId}{D1Object}(x)
}
\arguments{
\item{x}{D1Object}

\item{...}{(not yet used)}
}
\value{
the formatId
}
\description{
Get the FormatId of the DataObject
}
\section{Methods (by class)}{
\itemize{
\item \code{D1Object}: 
}}

