% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNode.R
\docType{methods}
\name{getMNode}
\alias{getMNode}
\alias{getMNode,CNode,character-method}
\title{Get a reference to a node based on its identifier}
\usage{
getMNode(cnode, nodeid, ...)

\S4method{getMNode}{CNode,character}(cnode, nodeid)
}
\arguments{
\item{cnode}{The coordinating node to query for its registered Member Nodes}

\item{nodeid}{The standard identifier string for this node}

\item{...}{(Not yet used)}
}
\value{
the Member Node as an MNode reference, or NULL if not found
}
\description{
Get a reference to a node based on its identifier
}
\seealso{
\code{\link[=CNode-class]{CNode}}{ class description.}
}

