% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Client.R
\name{getMNodeId}
\alias{getMNodeId}
\title{Get for the member node identifier associated with this client.
One Member Node can be associated with the client as the default to which
data and metadata are written.}
\usage{
getMNodeId(x)
}
\description{
Get for the member node identifier associated with this client.
One Member Node can be associated with the client as the default to which
data and metadata are written.
}

