% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Client.R
\docType{methods}
\name{getMNodeId}
\alias{getMNodeId}
\alias{getMNodeId,D1Client-method}
\title{Get the member node identifier associated with this D1Client object..}
\usage{
getMNodeId(x)

\S4method{getMNodeId}{D1Client}(x)
}
\arguments{
\item{x}{A D1Client object.}
}
\value{
The Member Node identifier  as a character vector
}
\description{
One Member Node can be associated with the client as the default to which
data and metadata are written.
}
\examples{
\dontrun{
cli <- D1Client("STAGING2", "urn:node:mnTestKNB")
mn <- getMNodeId(cli)
}
}
\seealso{
\code{\link[=D1Client-class]{D1Client}}{ class description.}
}

