% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getPackage}
\alias{getPackage}
\title{Download a DataPackage from the DataONE Cloud}
\usage{
getPackage(x, identifier, ...)
}
\arguments{
\item{x}{D1Client}

\item{identifier}{character: identifier of the ResourceMap}

\item{...}{(not yet used)}
}
\value{
the data package
}
\description{
Given a valid identifier for a ResourceMap, instantiate an R DataPackage
object containing all of the package members.
}

