% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Node.R
\docType{methods}
\name{getQueryEngineDescription}
\alias{getQueryEngineDescription}
\alias{getQueryEngineDescription,D1Node,character-method}
\title{Retrieve a description of a query engine from a coordinating node or member node}
\usage{
getQueryEngineDescription(node, queryEngineName)

\S4method{getQueryEngineDescription}{D1Node,character}(node, queryEngineName)
}
\arguments{
\item{node}{The CNode or MNode to query}

\item{queryEngine}{The query engine name to get a description for.}

\item{node}{The CNode or MNode instance to list the query engines for.}
}
\value{
list The query engine description

list Objects that met the search criteria
}
\description{
Retrieve a description of a query engine from a coordinating node or member node

Query a node for the list of query engines available on the node
}
\examples{
\dontrun{
cn <- CNode("SANDBOX")
engineDesc <- getQueryEngineDescription(cn, "solr")
cat(sprintf("Query engine version: \%s\\n", engineDesc$queryEngineVersion))
cat(sprintf("Query engine name: \%s\\n", engineDesc$name))
engineDesc <- getQueryEngineDescription(cn, "solr")
head(engineDesc$queryFields, n=3L)
}
}

