% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/D1Node.R
\name{getSystemMetadata}
\alias{getSystemMetadata}
\title{Get the metadata describing system properties associated with an object on this Node.}
\usage{
getSystemMetadata(node, pid, ...)
}
\arguments{
\item{node}{The Node instance from which the SystemMetadata will be downloaded}

\item{pid}{The identifier of the object}
}
\value{
SystemMetadata for the object
}
\description{
The SystemMetadata includes information about the identity, type, access control, and other system
level details about the object.
}
\details{
This operation acts as the 'public' anonymous user unless an X.509 certificate is
present in the default location of the file system, in which case the access will be authenticated.
}

