% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNode.R
\docType{methods}
\name{hasReservation}
\alias{hasReservation}
\alias{hasReservation,CNode-method}
\title{Checks to determine if the supplied subject is the owner of the reservation of id.}
\usage{
hasReservation(cnode, ...)

\S4method{hasReservation}{CNode}(cnode, pid, subject = as.character(NA))
}
\arguments{
\item{cnode}{A CNode instance.}

\item{...}{Additional parameters.}

\item{pid}{The identifier that is being checked for existing as a reserved identifier or is in use as 
an identifier for an existing object}

\item{subject}{The subject of the principal (user) that made the reservation. If not specified, then}
}
\value{
A logical value where TRUE means a reservation exists for the specified pid by the subject.
}
\description{
Checks to determine if the supplied subject is the owner of the reservation of id.
}
\seealso{
\code{\link[=CNode-class]{CNode}}{ class description.}
}

