% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AuthenticationManager.R
\docType{methods}
\name{isAuthExpired}
\alias{isAuthExpired}
\alias{isAuthExpired,AuthenticationManager-method}
\title{CHeck if the currently valid authentication method has reached the expiratin time.}
\usage{
isAuthExpired(x, ...)

\S4method{isAuthExpired}{AuthenticationManager}(x)
}
\arguments{
\item{x}{an AuthenticationManager instance}

\item{...}{(Not yet used.)}
}
\value{
A logical value: TRUE if authenentication has expired, FALSE if not.
}
\description{
CHeck if the currently valid authentication method has reached the expiratin time.
}
\section{Methods (by class)}{
\itemize{
\item \code{AuthenticationManager}: 
}}

