% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AuthenticationManager.R
\docType{methods}
\name{isAuthValid}
\alias{isAuthValid}
\alias{isAuthValid,AuthenticationManager,D1Node-method}
\title{Verify authentication for a member node.}
\usage{
isAuthValid(x, node, ...)

\S4method{isAuthValid}{AuthenticationManager,D1Node}(x, node)
}
\arguments{
\item{x}{a AuthenticationManager instance}
}
\value{
A logical value: TRUE if authentication is valid, false if not.
}
\description{
The currently used authentication methonism (either tokens or X.509 certificates)
is checked and verified for the specified member node. If an authentication token is availalbe
via the dataone package Session Configuration, it will be used.  However, authentication tokens can only be used for
DataONE v2.0 or higher nodes. X.509 certificates can be used with DataONE v1.0.0 or higher member nodes.
See the \emph{"dataone"} vignette \emph{"dataone-overview"} for more information on authentication.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = AuthenticationManager,node = D1Node}: 
}}

