% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CNode.R
\docType{methods}
\name{listFormats}
\alias{listFormats}
\alias{listFormats,CNode-method}
\title{list formats}
\usage{
listFormats(cnode, ...)

\S4method{listFormats}{CNode}(cnode)
}
\arguments{
\item{cnode}{a valid CNode object}
}
\value{
Returns a dataframe of all object formats registered in the DataONE Object Format Vocabulary.
}
\description{
list of all object formats registered in the DataONE Object Format Vocabulary.
}
\examples{
\dontrun{
cn <- CNode()
listFormats(cn)
}
}
\author{
hart
}

