% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CNode.R
\name{listNodes}
\alias{listNodes}
\title{Get the list of nodes associated with a CN}
\usage{
listNodes(cnode, ...)
}
\arguments{
\item{cnode}{The coordinating node to query for its registered Member Nodes}
}
\value{
the list of nodes in the DataONE CN environment
}
\description{
Get the list of nodes associated with a CN
}

