% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Node.R
\name{listObjects}
\alias{listObjects}
\title{Retrieve the list of objects that match the search parameters}
\usage{
listObjects(node, ...)
}
\arguments{
\item{node}{The Node instance from which the SystemMetadata will be downloaded}
}
\value{
list Objects that met the search criteria
}
\description{
Retrieve the list of objects that match the search parameters
}

