% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNode.R
\docType{class}
\name{MNode-class}
\alias{MNode-class}
\title{MNode is a class representing a Member Node repository within the DataONE federation of data repositories.}
\description{
MNode provides functions interacting with the a DataONE Member Node repository, which
is a repository that provides access for reading and writing data and metadata using the common
DataONE service API.  The MNode API includes functions for retrieving data and metadata based on its
unique persistent identifier (pid), as well as for creating, updating, and archiving these data and
metadata objects.
}
\details{
Methods that perform write operations (such as create and update) on the Member Node generally require
authentication. In the version 2.0 library and higher, these operations can utilize an 
'authentication_token' option to provide credentials for write operations in DataONE.
The authentication token is obtained from DataONE (see your profile on https://search.dataone.org).
See the \code{vignette("dataone-overview")} for details.
Alternatively, the version 1.0 approach of using an X.509 certificate in a default location of the file 
system can also be used. This certificate provides authentication credentials from 
CILogon \url{https://cilogon.org/?skin=DataONE}.  See \code{vignette("dataone-overview")} for details.
}
\section{Slots}{

\describe{
\item{\code{endpoint}}{The url to access node services, which is the baseURL plus the version string}
}}
\section{Methods}{

\itemize{
 \item{\code{\link{MNode}}}{: Create a MNode object representing a DataONE Member Node repository.}
 \item{\code{\link{create}}}{: Create an object on a Member Node.}
 \item{\code{\link{get}}}{: Get the bytes associated with an object on the Member Node}
 \item{\code{\link{getCapabilities}}}{: Get the node capabilities description, and store the information in the MNode.}
 \item{\code{\link{generateIdentifier}}}{: Get a unique identifier that is generated by the Member Node repository and guaranteed to be unique.}
 \item{\code{\link{getPackage}}}{: Download a data package from a member node.}
 \item{\code{\link{update}}}{: Update an object on a Member Node, by creating a new object that replaces an original.}
 \item{\code{\link{updateSystemMetadata}}}{: Update the system metadata associated with an object.}
}
}
\examples{
\dontrun{
cn <- CNode("STAGING2")
mn <- getMNode(cn, "urn:node:mnTestKNB")
mnid <- mn@identifier
newid <- generateIdentifier(mn, "UUID")
cm <- CertificateManager()
u <- showClientSubject(cm)
testdf <- data.frame(x=1:10,y=11:20)
csvfile <- paste(tempfile(), ".csv", sep="")
write.csv(testdf, csvfile, row.names=FALSE)
f <- "text/csv"
size <- file.info(csvfile)$size
sha1 <- digest(csvfile, algo="sha1", serialize=FALSE, file=TRUE)
sysmeta <- new("SystemMetadata", identifier=newid, formatId=f, size=size, submitter=u, 
    rightsHolder=u, checksum=sha1, originMemberNode=mnid, authoritativeMemberNode=mnid)
response <- create(mn, newid, csvfile, sysmeta)
}
}
\seealso{
\code{\link{dataone}}{ package description.}
}

