% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D1Node.R
\name{ping}
\alias{ping}
\title{Test if a node is online and accepting DataONE requests}
\usage{
ping(node)
}
\arguments{
\item{node}{The CNode or MNode to check}
}
\value{
logical A logical value set to TRUE if the node is up and FALSE if it is not
}
\description{
Test if a node is online and accepting DataONE requests
}

