% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CNode.R, R/D1Client.R
\docType{methods}
\name{reserveIdentifier}
\alias{reserveIdentifier}
\alias{reserveIdentifier,CNode,character-method}
\alias{reserveIdentifier,D1Client,character-method}
\title{Reserve a identifier that is unique in the DataONE network.}
\usage{
reserveIdentifier(x, id, ...)

\S4method{reserveIdentifier}{CNode,character}(x, id)

\S4method{reserveIdentifier}{D1Client,character}(x, id)
}
\arguments{
\item{x}{The coordinating node to query for its registered Member Nodes}

\item{id}{The identifier that is to be reserved.}

\item{...}{Additional parameters.}
}
\value{
The reserved pid if it was sucessfully reserved, otherwise NULL
}
\description{
Once a an identifier has been reserved, it and can not be used by any other user.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = CNode,id = character}: 

\item \code{x = D1Client,id = character}: 
}}
\seealso{
\code{\link[=CNode-class]{CNode}}{ class description.}
}

