% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CNode.R
\docType{methods}
\name{reserveIdentifier}
\alias{reserveIdentifier}
\alias{reserveIdentifier,CNode,character-method}
\title{serves an identifier that is unique and can not be used by any other sessions.}
\usage{
reserveIdentifier(x, id, ...)

\S4method{reserveIdentifier}{CNode,character}(x, id, quiet = TRUE)
}
\arguments{
\item{x}{The coordinating node to query for its registered Member Nodes}

\item{...}{Additional parameters.}

\item{quiet}{A logical value - if TRUE (the default) then informational messages are not printed.}

\item{pid}{The identifier that is to be reserved.}
}
\value{
The reserved pid if it was sucessfully reserved, otherwise NULL
}
\description{
serves an identifier that is unique and can not be used by any other sessions.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = CNode,id = character}: 
}}
\seealso{
\link{http://mule1.dataone.org/ArchitectureDocs-current/apis/CN_APIs.html#CNCore.reserveIdentifier}
}

