% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reserveIdentifier}
\alias{reserveIdentifier}
\title{Reserve an Identifier in the DataONE System
Reserve an identifier for future use in the DataONE System.}
\usage{
reserveIdentifier(x, id, ...)
}
\arguments{
\item{x}{: D1Client}

\item{id}{: identifier to reserve}

\item{...}{(not yet used)}
}
\value{
true if reserved
}
\description{
Reserve an Identifier in the DataONE System
Reserve an identifier for future use in the DataONE System.
}

