% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthenticationManager.R
\docType{methods}
\name{restoreAuth}
\alias{restoreAuth}
\alias{restoreAuth,AuthenticationManager-method}
\title{Restore authentication (after being disabled with \code{obscureAuth}).}
\usage{
restoreAuth(.Object)

\S4method{restoreAuth}{AuthenticationManager}(.Object)
}
\arguments{
\item{.Object}{An AuthenticationManager instance}
}
\value{
The expiration date for the current authentication mechanism being used.
}
\description{
Restore authentication (after being disabled with \code{obscureAuth}).
}

