% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AuthenticationManager.R
\docType{methods}
\name{showAuth}
\alias{showAuth}
\alias{showAuth,AuthenticationManager-method}
\title{Display all authentication information}
\usage{
showAuth(x, ...)

\S4method{showAuth}{AuthenticationManager}(x)
}
\arguments{
\item{x}{an Authentication instance}

\item{...}{(Not yet used.)}
}
\description{
Display all authentication information
}
\section{Methods (by class)}{
\itemize{
\item \code{AuthenticationManager}: 
}}

