% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MNode.R
\docType{methods}
\name{update}
\alias{update}
\alias{update,MNode,character-method}
\title{Update an object on a Member Node, by creating a new object that replaces an original.}
\usage{
update(mnode, pid, ...)

\S4method{update}{MNode,character}(mnode, pid, filepath, newpid, sysmeta)
}
\arguments{
\item{pid}{The identifier of the object to be updated}

\item{...}{(Not yet used.)}

\item{filepath}{the absolute file location of the object to be uploaded}

\item{newpid}{The identifier of the new object to be created}

\item{sysmeta}{a SystemMetadata instance describing properties of the object}

\item{node}{The MNode instance on which the object will be created}
}
\value{
XML describing the result of the operation, including the identifier if successful
}
\description{
This method provides the ability to update a data or metadata object to the Member Node
provided in the \code{'mnode'} parameter.  In DataONE, both the original object and the new object are
maintained, each with its own persistent identifier, and the 'obsoletes' field in the SystemMetadata is
used to reflect the fact that the new object replaces the old.  Both objects remain accessible.
}
\details{
This operation requires an X.509 certificate to be present in the default location of the file
system. This certificate provides authentication credentials from
CILogon \url{https://cilogon.org/?skin=DataONE}.  See \code{\link{{CertificateManager}}} for details.
}
\section{Methods (by class)}{
\itemize{
\item \code{mnode = MNode,pid = character}: 
}}
\seealso{
\url{http://mule1.dataone.org/ArchitectureDocs-current/apis/.html#MNStorage.update}
}

