% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MNode.R
\docType{methods}
\name{updateSystemMetadata}
\alias{updateSystemMetadata}
\alias{updateSystemMetadata,MNode,character,SystemMetadata-method}
\title{Update the system metadata associated with an object.}
\usage{
updateSystemMetadata(node, pid, sysmeta, ...)

\S4method{updateSystemMetadata}{MNode,character,SystemMetadata}(node, pid,
  sysmeta)
}
\arguments{
\item{node}{The MNode instance from which the SystemMetadata will be downloaded}

\item{pid}{The identifier of the object}

\item{sysmeta}{a SystemMetadata instance with updated information.}
}
\value{
A logical value, TRUE if the operation was sucessful, FALSE if there was an error.
}
\description{
A modified SytemMetadata object can be sent to DataONE that contains
updated information. This function allow updating of the system metadata without
updating the object that it describes, so that mutable attribures such as accessPolicy
can be updated easily.
}
\section{Methods (by class)}{
\itemize{
\item \code{node = MNode,pid = character,sysmeta = SystemMetadata}: 
}}
\seealso{
\url{http://mule1.dataone.org/ArchitectureDocs-current/apis/MN_APIs.html#MNStorage.updateSystemMetadata}
}

